fix: 가격 long 타입 전환 및 테스트 시 contextload api key bean 생성 문제 해결

+++ src/main/java/com/tripleseven/orderapi/dto/order/OrderViewDTO.java
@@ -10,13 +10,13 @@ public class OrderViewDTO {
     private final Long orderId;
     private final LocalDate orderDate;
     private final Long bookId;
-    private final int price;
-    private final int amount;
+    private final long price; // 가격
+    private final int amount; // 수량
     private final OrderStatus orderStatus;
     private final String ordererName;
     private final String recipientName;
 
-    public OrderViewDTO(Long orderId, LocalDate orderDate, Long bookId, int price, int amount,
+    public OrderViewDTO(Long orderId, LocalDate orderDate, Long bookId, long price, int amount,
                         OrderStatus orderStatus, String ordererName, String recipientName) {
         this.orderId = orderId;
         this.orderDate = orderDate;
