+++ src/main/java/com/tripleseven/orderapi/repository/ordergrouppointhistory/querydsl/QueryDslOrderGroupPointHistoryRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.domain.Pageable;
 
 public interface QueryDslOrderGroupPointHistoryRepository {
 
-    Integer findTotalAmountByOrderGroupId(Long orderGroupId, HistoryTypes historyTypes);
+    Long findTotalAmountByOrderGroupId(Long orderGroupId, HistoryTypes historyTypes);
 
     Page<UserPointHistoryDTO> findUserPointHistories(Long memberId, LocalDateTime startDate, LocalDateTime endDate, Pageable pageable);
 }
